requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
DateTime==4.9
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
zope.interface==5.5.2
